+++ src/ray.cpp
@@ -12,7 +12,7 @@ Ray::Ray(const Ray& r) {
 }
 
 // Value Constructor
-Ray::Ray(V3 offset, V3 slope, unsigned int depth) {
+Ray::Ray(V3 offset, V3 slope, unsigned short depth) {
     this->o = offset;
     this->s = slope;
     this->depth = depth;
